refactor(frontend): split mobile surface responsibilities
